part 4 done

+++ part4/controllers/blogs.js
@@ -14,8 +14,8 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!request.token ||!decodedToken.id) {
-    return response.status(401).json({ error: 'token missin or invalid'})
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid'})
   }
 
   const user = await User.findById(decodedToken.id)
@@ -37,8 +37,20 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog || blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(401).json({ error: 'operation not permitted'})
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  response.status(204).end()  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
